created the seperate modal for no thread and thread in lined_proposal card
